fonctionalité pour le prototype

+++ vyny/Point_Chauffeur/chauffeur_detect/views.py
@@ -28,8 +28,8 @@ from rest_framework_simplejwt.tokens import AccessToken
 from rest_framework.permissions import IsAuthenticated , AllowAny
 from rest_framework.authentication import TokenAuthentication
 import numpy as np
-from sklearn.metrics.pairwise import cosine_similarity
-
+from rest_framework.views import APIView
+from rest_framework_simplejwt.authentication import JWTAuthentication
 
 @api_view(['GET'])
 def hello_world(request):
@@ -120,32 +120,29 @@ def verifier_code(request):
 def add_chauffeur(request):
     item = ChauffeurSerializer(data=request.data)
     if item.is_valid():
+        numero_cni = request.data.get('numero_CNI')
+        if Chauffeurs.objects.filter(numero_CNI=numero_cni).exists():
+            return Response({"error": "Un chauffeur avec ce numéro CNI existe déjà"},status=status.HTTP_400_BAD_REQUEST)
+            
         # Récupérer l'image de la requête
         image_file = request.FILES.get('photo')
         if not image_file:
             return Response({"error": "Aucune image envoyée"}, status=400)
 
+        # Sauvegarder temporairement l'image
         filename = str(uuid.uuid4())
         temp_file_path = os.path.join(settings.MEDIA_ROOT, filename)
         with open(temp_file_path, 'wb') as temp_file:
             for chunk in image_file.chunks():
                 temp_file.write(chunk)
-        new_embedding = generate_emb(temp_file_path)
-        os.remove(temp_file_path) 
-        if new_embedding is None:
+        # Générer l'embedding à partir de l'image
+        embedding = generate_emb(temp_file_path)
+        os.remove(temp_file_path)  # Supprimer l'image temporaire après traitement
+        if embedding is None:
             return Response({"error": "Impossible d'extraire un visage de l'image"}, status=400)
-        else :
-            new_embedding = new_embedding.reshape(1, -1)
-            existing_chauffeurs = Chauffeurs.objects.all()
-        for chauffeur in existing_chauffeurs:
-            existing_embedding = np.frombuffer(chauffeur.embedding, dtype=np.float32)
-            existing_embedding = existing_embedding.reshape(1, -1) 
-            similarity = cosine_similarity(new_embedding, existing_embedding)
-            if similarity > 0.8: 
-                return Response({"error": "Ce visage est déjà enregistré"}, status=400)
-       
+        # Ajouter le chauffeur et son embedding
         chauffeur = item.save()
-        chauffeur.new_embedding = new_embedding.tobytes()
+        chauffeur.embedding = embedding.tobytes()  # Convertir l'embedding en format binaire
         chauffeur.save()
         return Response(ChauffeurSerializer(chauffeur).data, status=201)
     return Response(item.errors, status=400)
@@ -364,3 +361,10 @@ def update_recette(request, id):
         return Response({"SUCCESS": "UPDATE SUCCESS !"})
     else:
         return Response(serializerRecette.errors, status=status.HTTP_400_BAD_REQUEST)
+    
+# class VerifyTokenView(APIView):
+#     authentication_classes = [JWTAuthentication]
+#     permission_classes = [IsAuthenticated]
+
+#     def get(self, request):
+#         return Response({"message": "Token valide", "user": request.user.username}, status=200)
